Add DragonFlyBSD support
Add preliminary DragonFlyBSD support to ltsmaster. This is largely based
on copy/pasting the conditional compilation for FreeBSD, but changing
the guard from FreeBSD to DragonFlyBSD.

The only changes from FreeBSD are the pthread_cleanup_pop and
pthread_cleanup_push function names are different in pthread.d, and the
fstat structure is slightly different.