Fix profile timing measurements on AArch64.
We cannot use ldc.intrinsics.llvm_readcyclecounter because that is not an accurate
time counter (it is a counter of CPU cycles, where here we want a time clock).
Also, priviledged execution rights are needed to enable correct counting with
ldc.intrinsics.llvm_readcyclecounter on AArch64.
So instead, we use the general virtual counter register, and also read out the frequency of that counter.